+++ module_pwm/include/pwm_server.h
@@ -23,14 +23,18 @@
 #include <motor_control_interfaces.h>
 
 /**
+ * @brief Define updating period for general purpose pwm server while the ref_clk_freq is 100 MHz.
+ */
+#define GPWM_UPDATING_PERIOD          6667
+
+/**
  * @brief Define maximum possible values for general PWM server (which is able to
  * generate PWM pulses for up to 6 outputs.
  */
-/*
- * 6666  -> 15 kHz
- * 10000 -> 10 kHz
- */
-#define GPWM_MAX_VALUE      6667//10000
+#define GPWM_MAX_VALUE_15_kHz         6666
+#define GPWM_MAX_VALUE_30_kHz         3333
+#define GPWM_MAX_VALUE_100_kHz        1000
+
 /**
  * @brief Define maximum possible values for general PWM server (which is able to
  * generate PWM pulses for up to 6 outputs.
@@ -39,10 +43,10 @@
 
 /**
  * @brief PWM frequency in kHz.
- * Warning: So far, only GPWM_FRQ_15 is supported.
  */
-#define GPWM_FRQ_15               15 // PWM frequency in kHz.
-
+#define GPWM_FRQ_15               15 // PWM frequency in kHz
+#define GPWM_FRQ_30               30 // PWM frequency in kHz
+#define GPWM_FRQ_100              100// PWM frequency in kHz
 
 /**
  * @brief Structure type to define the ports to manage the FET-driver in your IFM SOMANET device (if applicable).
@@ -85,11 +89,12 @@ void pwm_config_general(PwmPortsGeneral &ports);
 
 /**
  * @brief Service to generate center-alligned PWM signals for 6 inverter outputs (2 power switch for each leg).
- * It recieves 6 pwm values through i_update_pwm interface. The commutation frequency is 16 kHz, and the deadtime is 3 us.
+ * It recieves 6 pwm values through i_update_pwm interface. The commutation frequency is 15 kHz.
  *
- * @param ports                 Structure type for PWM ports
+ * @param ports                 Structure type for general PWM ports
  * @param i_update_pwm          Interface to communicate with client and update the PWM values
  * @param freq_kHz              pwm frequency - kHz
+ * @param deadtime_ns           deadtime in [ns] (depends on board)
  *
  * @return void
  */